requirements.txt

django-widget-tweaks
django-recaptcha
